third_party/WebKit/LayoutTests/external/wpt/css/css-typed-om/stylevalue-subclasses/cssUnparsedValue-interface-expected.txt

This is a testharness.js-based test.
PASS CSSUnparsedValue interface: existence and properties of interface object
PASS CSSUnparsedValue interface object length
PASS CSSUnparsedValue interface object name
PASS CSSUnparsedValue interface: existence and properties of interface prototype object
PASS CSSUnparsedValue interface: existence and properties of interface prototype object's "constructor" property
PASS Testing Symbol.iterator property of iterable interface CSSUnparsedValue
FAIL Testing value iterable interface CSSUnparsedValue assert_equals: values function is not the same as Array one expected (undefined) undefined but got (function) function "function values() { [native code] }"
PASS CSSUnparsedValue interface: attribute length
FAIL CSSUnparsedValue interface: operation entries() assert_throws: calling operation with this = {} didn't throw TypeError function "function() {
            fn.apply(obj, args);
        }" did not throw
FAIL CSSUnparsedValue interface: operation keys() assert_throws: calling operation with this = {} didn't throw TypeError function "function() {
            fn.apply(obj, args);
        }" did not throw
FAIL CSSUnparsedValue interface: operation values() assert_throws: calling operation with this = {} didn't throw TypeError function "function() {
            fn.apply(obj, args);
        }" did not throw
PASS CSSUnparsedValue interface: operation forEach(function, any)
PASS CSSUnparsedValue must be primary interface of new CSSUnparsedValue()
PASS Stringification of new CSSUnparsedValue()
PASS CSSUnparsedValue interface: new CSSUnparsedValue() must inherit property "length" with the proper type
PASS CSSUnparsedValue interface: new CSSUnparsedValue() must inherit property "entries()" with the proper type
PASS CSSUnparsedValue interface: new CSSUnparsedValue() must inherit property "keys()" with the proper type
PASS CSSUnparsedValue interface: new CSSUnparsedValue() must inherit property "values()" with the proper type
PASS CSSUnparsedValue interface: new CSSUnparsedValue() must inherit property "forEach(function, any)" with the proper type
PASS CSSUnparsedValue interface: calling forEach(function, any) on new CSSUnparsedValue() with too few arguments must throw TypeError
PASS CSSUnparsedValue must be primary interface of new CSSUnparsedValue("foo")
PASS Stringification of new CSSUnparsedValue("foo")
PASS CSSUnparsedValue interface: new CSSUnparsedValue("foo") must inherit property "length" with the proper type
PASS CSSUnparsedValue interface: new CSSUnparsedValue("foo") must inherit property "entries()" with the proper type
PASS CSSUnparsedValue interface: new CSSUnparsedValue("foo") must inherit property "keys()" with the proper type
PASS CSSUnparsedValue interface: new CSSUnparsedValue("foo") must inherit property "values()" with the proper type
PASS CSSUnparsedValue interface: new CSSUnparsedValue("foo") must inherit property "forEach(function, any)" with the proper type
PASS CSSUnparsedValue interface: calling forEach(function, any) on new CSSUnparsedValue("foo") with too few arguments must throw TypeError
Harness: the test ran to completion.
